Module kb::bit::manipulation

Bit manipulation functions

Functions

is_exp2_minus_one
Determines if a bit vector is of the form 2ⁿ - 1 (or zero)

is_power_of_two
Determines if a bit vector is a power of two (or zero)

set_rightmost_zero
Sets the rightmost 0 in the bit vector

set_trailing_zeros
Sets the trailing 0s in the bit vector

unset_rightmost_one
Unsets the rightmost 1 in the bit vector

unset_trailing_ones
Unsets the trailing 1s in the bit vector