Function kb::math::gcd::euclid

pub fn euclid<N>(m: N, n: N) -> N where
    N: Unsigned + Copy + PartialOrd, 

Euclid's algorithm

Inputs: Two positive integers, m and n

Find the remainder of m and n, such that 0 ≤ r < n

If it's zero, terminate with n

Else, reduce by setting m ← n, n ← r and go back to step 1

Outputs: n, the greatest common divisor of m and n

Example

m = 119, n = 544
r ← 119, m ← 544, n = 119
r ← 68, m ← 119, n ← 68
r ← 51, m ← 68, n ← 51
r ← 17, m ← 51, n ← 17
r ← 0
terminate with 17

Proof

After step one, m = qn + r. If r is zero, m is a multiple of n, and n is the greatest common divisor of the two. If not, note that that a number that divides m and n must also divide r (because m - qn = r) and any number that divides n and r must divide m (because qn + r = m), so the set of common divisors for m and n is the same as that of n and r, as is the value of the greatest common divisor. Hence, step three does not change the answer to the original problem.

Invariants

If m < n originally, the quotient in step one will always be zero, so the two values will be swapped. It can be more efficient to do this in advance:

Ensure m ≥ n: if m < n, m ↔ n

Features

The algorithm satisfies five essential features:

Finiteness: After step one, r < n. If r is not zero, n decreases the next time step one is encountered. A decreasing sequence of positive integers will eventually terminate.

Definiteness: Undefined behaviour, such as division by zero, is not encountered because the inputs must be positive integers. After step one, r is a nonnegative integer that must be nonzero to encounter step three, so m and n will remain positive integers as required.

Inputs: m and n, where m ∈ ℕ⁺ and n ∈ ℕ⁺

Outputs: n in step two, the greatest common divisor of the inputs

Effectiveness: All operations are sufficiently basic such that they are exact and can be done in constant time.

Computational method

The following is functional representation of the algorithm in Haskell.

input(m, n) = normal(m, n, 0, 1)
terminate(n) = n

normal(m, n, r, 1) = normal(m, n, (rem m n), 2)
normal(m, n, r, 2) = if r == 0 then terminate(n) else normal(m, n, r, 3)
normal(m, n, p, 3) = normal(n, p, p, 1)

Tests

assert_eq!(euclid(119_u32, 544), 17);
assert_eq!(euclid(252_u32, 105), 21);
assert_eq!(euclid(6_u32, 3), 3);
assert_eq!(euclid(6_u32, 4), 2);
assert_eq!(euclid(2166_u32, 6099), 57);