Module kb::search::linear

Linear search

Searches for an element in O(n) time.

On average, assuming an equal probability of occurrence of 1/n, the average number of elements that need to be checked is

1/n * (1 + 2 + ... + n)

which is equivalent to (by derivation from partial sums)

(n + 1) / 2

In the worst case, the target element does not exist in the array, meaning n checks have been made.

Complexity

Average case: O(n) because n checks have to be made, ignoring insignificant terms.

Worst case: O(n) because n checks have to be made.

Functions

linear
My own implementation of CLRS exercise 2.1-3