Function kb::search::linear::linear

pub fn linear<T>(target: &T, values: &[T]) -> Option<usize> where
    T: PartialEq, 

My own implementation of CLRS exercise 2.1-3

Iterates through a slice until a target value is found. If the value is found, Some(index) is returned, otherwise None is returned.

Loop invariant

The target value must have been found, or we keep searching until we have compared every element.

Initialization: The value has not been found, so we continue the search.

Maintenance: For each iteration, we check if the value has been found. If it has, we return the index at which it appears, otherwise we proceed to the next iteration.

Termination: If we found the value, its index is returned. If not, we have exhausted the search.