Module kb::sort::insertion

Insertion sort

Sorts elements in O(n²) time. Imagine having a set of cards face down on a table, picking them up in their original unsorted order, and then inserting them into the correct order in hand.

Best case: O(n) for when the array is already sorted. The inner loop never executes because the key is always greater than or equal to the element before it.

Worst case: O(n²) for when the array is reverse sorted. The key is always less than or equal to the element before it. If it is less than, elements must be shifted. O(n) elements must be shifted for the last element, if it is distinct. This happens in the inner loop.

Example

Say we have the following array:

[31, 41, 59, 26, 41, 58]

Insertion sort proceeds as follows:

+----+----+----+----+----+----+
|    |    |    |    |    |    |
| 31 | 41 | 59 | 26 | 41 | 58 |
|    |    |    |    |    |    |
+----++--++----+----+----+----+
      ^  |
      +--+

+----+----+----+----+----+----+
|    |    |    |    |    |    |
| 31 | 41 | 59 | 26 | 41 | 58 |
|    |    |    |    |    |    |
+----+----++--++----+----+----+
           ^  |
           +--+

+----+----+----+----+----+----+
|    |    |    |    |    |    |
| 31 | 41 | 59 | 26 | 41 | 58 |
|    |    |    |    |    |    |
++--+++--+++--+++-+-+----+----+
 ^  | ^  | ^  | ^ |
 |  +-+  +-+  +-+ |
 +----------------+

+----+----+----+----+----+----+
|    |    |    |    |    |    |
| 26   31   41   59   41   58 |
|    |    |    |    |    |    |
+----+----+----++--+++-+-+----+
                ^  | ^ |
                |  +-+ |
                +------+

+----+----+----+----+----+----+
|    |    |    |    |    |    |
| 26   31   41   41   59   58 |
|    |    |    |    |    |    |
+----+----+----+----++--+++-+-+
                     ^  | ^ |
                     |  +-+ |
                     +------+

+----+----+----+----+----+----+
|    |    |    |    |    |    |
| 26   31   41   41   58   59 |
|    |    |    |    |    |    |
+----+----+----+----+----+----+

Functions

clrs
The solution presented in CLRS

shepmaster
The solution found in this Code Review response