Function kb::sort::insertion::clrs

pub fn clrs<T>(arr: &mut [T]) where
    T: Copy + PartialOrd, 

The solution presented in CLRS

This solution matches the one presented in CLRS with one exception: i is set to j initially, rather than j - 1, and the inner loop only executes when i if not indexing the first element. This prevents i from ever being negative, which causes overflow if i is a usize (what it logically should be).

The outer loop starts from the second element. The inner loop repeatedly shifts elements on its left that are greater than it one space to their right, until the correct position for the key is found. The key is then written in that position.

Loop invariant

At the start of each iteration of the for loop, the slice arr[0..j] consists of the elements originally in arr[0..j], but in sorted order. This property must hold for initialization, maintenance and termination.

Initialization: This is the slice [0..1], a slice of one element, which is sorted.

Maintenance: The for loop moves elements larger than arr[j] one space to the right repeatedly until the correct space for arr[j] is found, at which point it inserts arr[j], and the slice arr[0..j] is sorted.

Termination: Upon termination, j == n, so the slice arr[0..j] is the entire array, which is sorted.