Function kb::sort::merge::clrs_merge

pub fn clrs_merge(values: &mut [f64], p: usize, q: usize, r: usize)

CLRS version of the merge operation used by merge sort

Merges two sorted subsequences in O(n) time, where n = r - p.

This implementation includes a sentinel value to simplify the implementation. It limits the sort to types that have a logical maximum value.

Loop invariant

At the start of the merge operation loop, values[p..k] contains the k - p smallest elements of left[0..n1] and right[0..n2], in sorted order. Also, left[i] and right[j] are the smallest elements of their slices that have yet to be copied back into values.

Initialization: To begin with, k = p, so values[p..k] is empty, which contains the zero smallest elements of left and right. Also, i and j point to the first elements of their slices, which are the smallest elements because the slices are sorted.

Maintenance: If left[i] is less than right[j], left[i] will be copied, meaning values[p..k + 1] will now contain the k - p + 1 smallest elements. k and i are both incremented, reestablishing the loop invariant for the next iteration. This process is symmetric if right holds the next smallest element.

Termination: Upon termination, k = r, so values[p..k] is values[p..r], and holds the k - p smallest elements in left and right. left and right are both pointing to their sentinel values, which are the smallest values in their respective slices.

Example

Imagine having two sorted sets of cards face up on a table, with the smallest card on top. We want to merge the two piles into one pile. We do this by taking the smaller of the two visible cards and placing them in a new pile. Once one of the input piles is empty, we take all of the remaining cards in the remaining pile and put them in the output pile.

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 2 | 4 | 5 | 7 | 1 | 2 | 3 | 6 |
          |   |   |   |   |   |   |   |   |
          +--++---+---+---+---+---+---+---+
           k ^
             +------------------+
                                |
  +---+---+---+---+---+       +-+-+---+---+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
    i                           j

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 4 | 5 | 7 | 1 | 2 | 3 | 6 |
          |   |   |   |   |   |   |   |   |
          +---+--++---+---+---+---+---+---+
               k ^
    +------------+
    |
  +-+-+---+---+---+---+       +---+---+---+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
    i                               j

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 2 | 5 | 7 | 1 | 2 | 3 | 6 |
          |   |   |   |   |   |   |   |   |
          +---+---+--++---+---+---+---+---+
                   k ^
                     +--------------+
                                    |
  +---+---+---+---+---+       +---+-+-+---+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
        i                           j

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 2 | 2 | 7 | 1 | 2 | 3 | 6 |
          |   |   |   |   |   |   |   |   |
          +---+---+---+--++---+---+---+---+
                       k ^
                         +--------------+
                                        |
  +---+---+---+---+---+       +---+---+-+-+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
        i                               j

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 2 | 2 | 3 | 1 | 2 | 3 | 6 |
          |   |   |   |   |   |   |   |   |
          +---+---+---+---+--++---+---+---+
                           k ^
        +--------------------+
        |
  +---+-+-+---+---+---+       +---+---+---+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
        i                                   j

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 2 | 2 | 3 | 4 | 2 | 3 | 6 |
          |   |   |   |   |   |   |   |   |
          +---+---+---+---+---+--++---+---+
                               k ^
            +--------------------+
            |
  +---+---+-+-+---+---+       +---+---+---+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
            i                               j

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 2 | 2 | 3 | 4 | 5 | 3 | 6 |
          |   |   |   |   |   |   |   |   |
          +---+---+---+---+---+---+--++---+
                                   k ^
                                     +------+
                                            |
  +---+---+---+---+---+       +---+---+---+-+-+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
                i                           j

          +---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 2 | 2 | 3 | 4 | 5 | 6 | 6 |
          |   |   |   |   |   |   |   |   |
          +---+---+---+---+---+---+---+--++
                                       k ^
                +------------------------+
                |
  +---+---+---+-+-+---+       +---+---+---+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
                i                               j

          +---+---+---+---+---+---+---+---+---+
          |   |   |   |   |   |   |   |   |
        V | 1 | 2 | 2 | 3 | 4 | 5 | 6 | 7 |
          |   |   |   |   |   |   |   |   |
          +---+---+---+---+---+---+---+---+---+
                                            k

  +---+---+---+---+---+       +---+---+---+---+---+
  |   |   |   |   |   |       |   |   |   |   |   |
L | 2 | 4 | 5 | 7 | ∞ |     R | 1 | 2 | 3 | 6 | ∞ |
  |   |   |   |   |   |       |   |   |   |   |   |
  +---+---+---+---+---+       +---+---+---+---+---+
                    i                           j