Module kb::sort::selection

Selection sort

Loop invariant

The slice [0..i] is always in sorted order.

Initialization: i is 0, so the slice is empty, which is sorted.

Maintenance: At the end of each iteration of the outer loop, min will be pointing to the minimum value in the slice [i..n - 1]. This value is swapped with i, so the new value at i is always smaller than or equal to all subsequent values, and [0..i] is still sorted.

Termination: i is equal to n - 1 and there have been n - 1 swaps. The slice [0..n - 1] contains the smallest n - 1 elements, therefore the final element is already in the correct place. [0..n] is now sorted.

Complexity

Best case: O(n²) because the inner loop has to iterate through i + 1..n values to search for the minimum, regardless of how sorted the array already is.

Word case: O(n²) because the inner loop has to iterate through i + 1..n values to search for the minimum, regardless of how sorted the array already is.

Functions

selection
My own implementation of CLRS exercise 2.2-2